requirements.txt
fastapi~=0.111.0
pydantic~=2.7.4
python-multipart==0.0.5
expiringdict==1.2.2
rich==13.4.2
openai==1.6.1
httpx==0.27.0


requests~=2.31.0
pillow~=10.3.0